Manually encode semicolons in query strings for vivaldi
It seems that Vivaldi does not encode them properly, so we now do it for
them.  The issue only seems to appear when running in the docker
container using gunicorn, when running locally with only uvicorn the
same problem does not seem to happen.

This fix means we now likely break some part of the URL specs, as
semicolons are reserved characters, but we should not break anything, as
we are just forcing the encoding of semicolons.

Despite the `.decode("latin-1")` emojis and other Unicode characters
still work, so that should be fine as well.