Fix "remove field" buttons in the table header
These were broken because first the mouse click handler for `.filter`
was overwriting the click, which lead to `&undefined=✖` to be appended
to the url.  The second issue was that even without the handler breaking
things the `&fields` parameter would be appended multiple times.  Now we
overwrite it in a copy of the query and the parameter only appears once.