Fix histogram tooltips on chromium
The SVG 1.1 standard does not support linebreaks [1] and chromium does
not implement them as well.  Firefox does, which is why I have not done
this for so long.

Overall, the code looks nicer like this and it's clearer what actual
rendered SVG will look like from the template code now.

[1]: https://stackoverflow.com/questions/16701522/how-to-linebreak-an-svg-text-within-javascript