Update all dependencies
The workflow for this is quite strange, at least `pip` only provides
a `--upgrade <dependencies...>` flag, but there does not seem to be
an automatic update workflow, especially taking into account the
potential conflicting dependencies of other packages.

But this seems to work now, all dependencies were updated and we can
drop `gunicorn` which is not used anymore since the switch to the
"more Kubernetes native" deployment approach.