Allow specifying multiple hosts per ElasticSearch endpoint
The ElasticSearch clients have always allowed this, but we haven't used it.
This is more flexible and should work in more situations.

This is a **breaking change** to the config, but as the API and everything
else is still the same this will be released as a minor update.  (We would
have the option to support both syntaxes, but the config is easy to change
so that seems like unnecessary complexity, especially with the... small
number of users of this project that exist.)