Add buckets for values without the aggregated field
E.g. when aggregating on `status.code` not all logs may have that field.
Before they would not be represented in the histogram and result in
misleading graphs, now they are shown as grey boxes on top of the ones
that are aggregated on.

This happened a couple times internally and should help point out
weird values and prevent confusion when expected or assumed fields
are missing.